pyproject.toml
[project]
name = "ai-study"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-generativeai>=0.8.6",
    "pygame-ce>=2.5.7",
    "python-dotenv>=1.2.2",
]
